requirements.txt
plotly_express
pandas
streamlit
altair
vega_datasets